Factor out qt_octree_info.py file
